Fix behaviour of read() if less data is available than requested.
read() shall read at most n bytes from a FD. If less than n bytes
are available it shall read them and return the number of bytes
actually read.
If more data is requested than there is available, instead of clearing
the destination buffer and returning an error, a deliver the available
data.

Signed-off-by: Florian Achleitner <florian.achleitner@student.tugraz.at>