Fix fstat()
fstat() returned error -1 on every call and logged "Failed to read
file data!". The reason is that the file is lseeked to the end of the
last block, before trying to read it, which fails, consequently.
Correct the lseek-position to the beginning of the last block to read
the required metadata for the encrypted file.

Signed-off-by: Florian Achleitner <florian.achleitner@student.tugraz.at>