Improve mmu translation function error reporting
When processing a software-requested translation, in case of a
translation fault, the mmu returns error bits in the
physical address register (PAR).
Detect this errors according to the spec in the reference manual and
deliver the error bits.
Preserve the existing easy-to-use function v_to_p.
Remove the duplicate virt_to_phys