Add TLS parsing to elf loader
Detect TLS program header and record it's contents for further use, when
parsing the elf file.

TLS (thread local storage) follows the local exec TLS model. The elf
file contains a TLS template which has to be copied into a seperate
instance for every thread. The template is designated by the TLS program
header. The TLS section is part of a LOAD section. Therefore, the
template is already loaded into memory, we only need to record where it
is.